market_intel/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
drf-yasg==1.21.7
setuptools>=80.9.0
selenium==4.27.1
pandas==2.2.3
pyarrow==17.0.0
numpy==2.1.3
scikit-learn==1.6.0
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2
webdriver-manager==4.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
twikit>=2.0.0
requests>=2.31.0
colorlog==6.8.2

